restructured tables, moved fields table out of contentTypesTable component
