+++ src/components/Tables/Field/MissingFieldsTable.tsx
@@ -1,0 +1,3 @@
+export function MissingFieldsTable() {
+  return <div />;
+}
